refactor : MVC 적용
