+++ src/main/java/lotto/Application.java
@@ -1,5 +1,8 @@
 package lotto;
 
+import lotto.controller.InstanceManager;
+import lotto.view.LottoGame;
+
 public class Application {
     public static void main(String[] args) {
         InstanceManager instanceManager = new InstanceManager();
