+++ src/main/java/lotto/model/RankEnum.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 public enum RankEnum {
     FIRST(6, 2_000_000_000),
