+++ src/main/java/lotto/model/WinningResult.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 
 import java.util.EnumMap;
