+++ src/main/java/lotto/view/ConsoleInput.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import java.util.HashSet;
